First draft.
Downloads and installs Java 8 in a user directory.
Downloads and installs ant and git.
Clones jSuneido and suneido.js.
Build jSuneido.
Creates an alias to start jsuneido.
Note that jSuneido doesn't seem to work if you try to put the database on
the /vagrant device. The alias does cd /home/vagrant before the commend.