Partly finished. :-)
Added the upstart script to start.
Changed the start command into a script, and added the startup for the
jsuneido server, but I haven't tested yet. I'm not sure that I can run
jsuneido as a server, and also pass in the command to start the web server.

This checkin is broken, because I don't have somewhere to get stdlib.su from.
Also, it simply doesn't load stdlib when I put it there manually. I haven't
yet tried to figure out why.