Closer to full startup.
The rest of the command line is a command when starting jsuneido, so
I now have start, init and stop working. The upstart scripts don't yet
work due to permissions on the log files.

I still don't know where to get stdlib from.

I need to set the environment variables so that the upstart script will
work. It doesn't know about HOME or any other stuff.