Updates stdlib.
Added the update of stdlib. Took a lot of experimentation. APM said don't
update stdlib on a running server, and that seemed to be correct. Updating
the server runs the Sundeido shell after, so I took stdin from /dev/null
and it quits on EOF. Not the most elegant.

An aside: Don't know updating is supposed to work in the libview.html window
if we shouldn't update stdlib in a running server?